:ambulance: new approach: an user iap account_token will be generated in the proxy server; One user will have one iap account_token regardless of how many clients (i.e. odoo instances with the module) he has.

+++ odoo_backup_sh/controllers/main.py
@@ -12,21 +12,30 @@ from contextlib import closing
 from datetime import datetime, timedelta
 
 import odoo
-from odoo import fields, http, _
+from odoo import exceptions, fields, http, _
 from odoo.exceptions import UserError
 from odoo.http import request
 from odoo.service import db
 from odoo.sql_db import db_connect
-from odoo.tools import DEFAULT_SERVER_DATETIME_FORMAT, DEFAULT_SERVER_DATE_FORMAT
+from odoo.tools import config, DEFAULT_SERVER_DATETIME_FORMAT, DEFAULT_SERVER_DATE_FORMAT
 from odoo.tools.misc import str2bool
 from odoo.addons import web
 from odoo.addons.web.controllers.main import DBNAME_PATTERN
 
 try:
+    import configparser as ConfigParser
+except ImportError:
+    import ConfigParser
+
+try:
+    import boto3
+    import botocore
     from pretty_bad_protocol import gnupg
 except ImportError as err:
     logging.getLogger(__name__).debug(err)
 
+_logger = logging.getLogger(__name__)
+config_parser = ConfigParser.ConfigParser()
 path = os.path.realpath(os.path.join(os.path.dirname(__file__), '..', 'views'))
 loader = jinja2.FileSystemLoader(path)
 env = jinja2.Environment(loader=loader, autoescape=True)
@@ -46,7 +55,7 @@ class BackupDatabase(web.controllers.main.Database):
                 '<a role="button" data-toggle="modal" data-target=".o_database_restore" class="btn btn-link">'
             )
         if place_for_backup_button != -1:
-            d['list_db'] = odoo.tools.config['list_db']
+            d['list_db'] = config['list_db']
             d['databases'] = []
             try:
                 d['databases'] = http.db_list()
@@ -62,13 +71,13 @@ class BackupDatabase(web.controllers.main.Database):
 class BackupController(http.Controller):
 
     @classmethod
-    def get_cloud_params(cls, redirect=None):
-        cloud_params = request.env['odoo_backup_sh.odoo_tools_config'].get_values('options', [
+    def get_cloud_params(cls, redirect=None, call_from='backend'):
+        cloud_params = cls.get_config_values('options', [
             'odoo_backup_user_key', 'amazon_bucket_name', 'amazon_access_key_id', 'amazon_secret_access_key',
             'odoo_oauth_uid'])
         if None in cloud_params.values():
             if redirect:
-                cloud_params = cls.update_cloud_params(cloud_params, redirect)
+                cloud_params = cls.update_cloud_params(cloud_params, redirect, call_from)
             else:
                 # This case occurs when we cannot pass the redirect param, ex. the request was sent by a cron
                 raise UserError(_("You don't have credentials to access to cloud storage. "
@@ -76,17 +85,16 @@ class BackupController(http.Controller):
         return cloud_params
 
     @classmethod
-    def update_cloud_params(cls, cloud_params, redirect):
+    def update_cloud_params(cls, cloud_params, redirect, call_from):
         user_key = cloud_params['odoo_backup_user_key']
         if not user_key:
             user_key = ''.join(random.choice(string.hexdigits) for _ in range(30))
-            request.env['odoo_backup_sh.odoo_tools_config'].set_values('options', {'odoo_backup_user_key': user_key})
+            cls.set_config_values('options', {'odoo_backup_user_key': user_key})
         cloud_params = requests.get(BACKUP_SERVICE_ENDPOINT + '/get_cloud_params', params={
             'user_key': user_key,
             'redirect': redirect,
-            'account_token': request.env['iap.account'].sudo().get('odoo_backup_sh').account_token,
         }).json()
-        if 'insufficient_credit_error' in cloud_params:
+        if 'insufficient_credit_error' in cloud_params and call_from == 'backend':
             existing_msg = request.env['odoo_backup_sh.notification'].sudo().search(
                 [('type', '=', 'insufficient_credits'), ('is_read', '=', False)])
             notification_vals = {
@@ -99,42 +107,27 @@ class BackupController(http.Controller):
             else:
                 request.env['odoo_backup_sh.notification'].sudo().create(notification_vals)
         elif 'auth_link' not in cloud_params:
-            if 'date_first_payment' in cloud_params:
-                if not request.env['ir.cron'].sudo().search([('name', '=', 'Odoo-backup.sh: make next payment')]):
-                    request.env['ir.cron'].sudo().create({
-                        'name': 'Odoo-backup.sh: make next payment',
-                        'model_id': request.env.ref('odoo_backup_sh.model_odoo_backup_sh_config').id,
-                        'state': 'code',
-                        'code': 'model.make_payment()',
-                        'interval_number': 30,
-                        'interval_type': 'days',
-                        'numbercall': -1,
-                        'nextcall': (datetime.strptime(
-                            cloud_params['date_first_payment'], DEFAULT_SERVER_DATETIME_FORMAT) + timedelta(days=30)),
-                        'doall': True,
-                        'active': True,
-                    })
-                del cloud_params['date_first_payment']
-            request.env['odoo_backup_sh.odoo_tools_config'].set_values('options', cloud_params)
-            request.env['odoo_backup_sh.notification'].sudo().search(
-                [('type', 'in', ['insufficient_credits', 'forecast_insufficient_credits']), ('is_read', '=', False)]
-            ).unlink()
+            cls.set_config_values('options', cloud_params)
+            if call_from == 'backend':
+                request.env['odoo_backup_sh.notification'].sudo().search(
+                    [('type', 'in', ['insufficient_credits', 'forecast_insufficient_credits']), ('is_read', '=', False)]
+                ).unlink()
         return cloud_params
 
     @http.route('/web/database/backups', type='http', auth="none")
     def backup_list(self):
-        cloud_params = self.get_cloud_params(request.httprequest.url)
+        cloud_params = self.get_cloud_params(request.httprequest.url, call_from='frontend')
         page_values = {
             'backup_list': [],
             'pattern': DBNAME_PATTERN,
-            'insecure': odoo.tools.config.verify_admin_password('admin')
+            'insecure': config.verify_admin_password('admin')
         }
         if 'auth_link' in cloud_params:
             return "<html><head><script>window.location.href = '%s';</script></head></html>" % cloud_params['auth_link']
         elif 'insufficient_credit_error' in cloud_params:
             page_values['error'] = cloud_params['insufficient_credit_error']
             return env.get_template("backup_list.html").render(page_values)
-        backup_list = request.env['odoo_backup_sh.cloud_storage'].get_backup_list(cloud_params)
+        backup_list = BackupCloudStorage.get_backup_list(cloud_params)
         if 'reload_page' in backup_list:
             page_values['error'] = 'Something went wrong. Please refresh the page.'
             return env.get_template("backup_list.html").render(page_values)
@@ -143,12 +136,12 @@ class BackupController(http.Controller):
 
     @http.route('/web/database/restore_via_odoo_backup_sh', type='http', auth="none", methods=['POST'], csrf=False)
     def restore_via_odoo_backup_sh(self, master_pwd, backup_file_name, name, copy=False):
-        cloud_params = self.get_cloud_params(request.httprequest.url)
-        backup_object = request.env['odoo_backup_sh.cloud_storage'].get_object(cloud_params, backup_file_name)
+        cloud_params = self.get_cloud_params(request.httprequest.url, call_from='frontend')
+        backup_object = BackupCloudStorage.get_object(cloud_params, backup_file_name)
         backup_file = tempfile.NamedTemporaryFile()
         backup_file.write(backup_object['Body'].read())
         if backup_file_name.split('|')[0][-4:] == '.enc':
-            passphrase = request.env['odoo_backup_sh.odoo_tools_config'].get_values(
+            passphrase = self.get_config_values(
                 'options', ['odoo_backup_encryption_password'])['odoo_backup_encryption_password']
             if not passphrase:
                 raise UserError(_(
@@ -184,7 +177,7 @@ class BackupController(http.Controller):
         date_list = [date_month_before + timedelta(days=x) for x in range(30)]
         last_month_domain = [('date', '>=', datetime.strftime(date_list[0], DEFAULT_SERVER_DATE_FORMAT))]
         usage_values = {
-            datetime.strptime(r.date, DEFAULT_SERVER_DATE_FORMAT).date(): r.total_used_remote_storage for r in
+            r.date: r.total_used_remote_storage for r in
             request.env['odoo_backup_sh.remote_storage'].search(last_month_domain).sorted(key='date')
         }
         for date in date_list:
@@ -211,8 +204,7 @@ class BackupController(http.Controller):
             for backup in request.env['odoo_backup_sh.backup_info'].search([
                     ('database', '=', b_config['database']),
                     ('upload_datetime', '>=', datetime.strftime(last_week_dates[0], DEFAULT_SERVER_DATETIME_FORMAT))]):
-                graph_values[datetime.strptime(backup.upload_datetime, DEFAULT_SERVER_DATETIME_FORMAT).date()] +=\
-                    backup.backup_size
+                graph_values[backup.upload_datetime.date()] += backup.backup_size
             b_config['graph'] = [{
                 'key': 'Backups of Last 7 Days',
                 'values': [{
@@ -238,3 +230,86 @@ class BackupController(http.Controller):
                 [('is_read', '=', False)], ['id', 'date_create', 'message'])
         })
         return dashboard_data
+
+    @classmethod
+    def get_config_values(cls, section, options_list):
+        """
+        :return dict: option_name: value
+        """
+        config_parser.read(config.rcfile)
+        result = {}
+        for option in options_list:
+            result[option] = config_parser.get(section, option, fallback=None)
+        return result
+
+    @classmethod
+    def set_config_values(cls, section, options_dict):
+        for option, value in options_dict.items():
+            config_parser.set(section, option, value)
+        with open(config.rcfile, 'w') as configfile:
+            config_parser.write(configfile)
+
+    @classmethod
+    def remove_config_options(cls, section, options_list):
+        config_parser.read(config.rcfile)
+        for option in options_list:
+            config_parser.remove_option(section, option)
+        with open(config.rcfile, 'w') as configfile:
+            config_parser.write(configfile)
+
+
+def access_control(origin_method):
+    def wrapped(self, *args, **kwargs):
+        try:
+            return origin_method(self, *args, **kwargs)
+        except botocore.exceptions.ClientError as client_error:
+            if client_error.response['Error']['Code'] == 'InvalidAccessKeyId':
+                BackupController.remove_config_options('options', ['amazon_access_key_id', 'amazon_secret_access_key'])
+                return {'reload_page': True}
+            else:
+                raise exceptions.ValidationError(_(
+                    "Amazon Web Services error: " + client_error.response['Error']['Message']))
+    return wrapped
+
+
+class BackupCloudStorage(http.Controller):
+
+    @classmethod
+    def get_amazon_s3_client(cls, cloud_params):
+        s3_client = boto3.client('s3', aws_access_key_id=cloud_params['amazon_access_key_id'],
+                                 aws_secret_access_key=cloud_params['amazon_secret_access_key'])
+        return s3_client
+
+    @classmethod
+    @access_control
+    def get_backup_list(cls, cloud_params):
+        amazon_s3_client = cls.get_amazon_s3_client(cloud_params)
+        user_dir_name = '%s/' % cloud_params['odoo_oauth_uid']
+        list_objects = amazon_s3_client.list_objects_v2(
+            Bucket=cloud_params['amazon_bucket_name'], Prefix=user_dir_name, Delimiter='/')
+        return {'backup_list': [
+            obj['Key'][len(user_dir_name):] for obj in list_objects.get('Contents', {}) if obj.get('Size')]}
+
+    @classmethod
+    @access_control
+    def get_object(cls, cloud_params, obj_name):
+        amazon_s3_client = cls.get_amazon_s3_client(cloud_params)
+        object_path = '%s/%s' % (cloud_params['odoo_oauth_uid'], obj_name)
+        return amazon_s3_client.get_object(Bucket=cloud_params['amazon_bucket_name'], Key=object_path)
+
+    @classmethod
+    @access_control
+    def put_object(cls, cloud_params, obj, obj_path):
+        amazon_s3_client = cls.get_amazon_s3_client(cloud_params)
+        amazon_s3_client.put_object(Body=obj, Bucket=cloud_params['amazon_bucket_name'], Key=obj_path)
+        _logger.info('Following backup object have been put in the remote storage: %s' % obj_path)
+
+    @classmethod
+    @access_control
+    def delete_objects(cls, cloud_params, objs):
+        amazon_s3_client = cls.get_amazon_s3_client(cloud_params)
+        amazon_s3_client.delete_objects(
+            Bucket=cloud_params['amazon_bucket_name'], Delete={'Objects': objs})
+        objects_names = [obj['Key'] for obj in objs]
+        _logger.info(
+            'Following backup objects have been deleted from the remote storage: %s' % ', '.join(objects_names))
